requirements.txt
absl-py==2.0.0
aiofiles==22.1.0
aiosqlite==0.19.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.1.0
Babel==2.13.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach==6.0.0
brotlipy==0.7.0
cached-property==1.5.2
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.13.0
conda-build==3.21.9
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649105789509/work
contextlib2==21.6.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652083456434/work
cycler==0.11.0
debugpy==1.7.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.1.3
ez-setup==0.9
fastjsonschema==2.18.1
filelock @ file:///opt/conda/conda-bld/filelock_1647002191454/work
fonttools==4.38.0
fqdn==1.5.1
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
gpytorch==1.6.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython @ file:///tmp/abs_94gruux8u8/croots/recipe/ipython_1659529858706/work
ipython-genutils==0.2.0
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.5.3
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1651563629415/work
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkginfo @ file:///tmp/build/80754af9/pkginfo_1643162084911/work
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
psutil @ file:///tmp/build/80754af9/psutil_1612298016854/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==12.0.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
pyzmq==24.0.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post10
sniffio==1.3.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.12.1
torchtext==0.13.1
torchvision==0.13.1
tornado==6.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.9.0
typing==3.7.4.3
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
uri-template==1.3.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.15.0
